backend/requirements.txt
# Core
fastapi==0.110.1
uvicorn[standard]==0.29.0

# Bases de datos (para luego conectar Supabase o PostgreSQL directo)
asyncpg==0.29.0
sqlalchemy==2.0.30

# Autenticación segura
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Cliente HTTP (para conectar APIs externas como Supabase REST o Disqus)
httpx==0.27.0

# Manejo de variables de entorno
python-dotenv==1.0.1

# CORS (por si necesitas exponer APIs para Astro Frontend)
fastapi[all]==0.110.1

# Validaciones avanzadas
pydantic==2.7.1

# Email con SendGrid (SMTP o API)
sendgrid==6.10.0
